move : 계층 이동 및 불필요한 파일 제거
