+++ src/main/java/dnd/myOcean/service/sign/KakaoMemberDetailService.java
@@ -1,6 +1,7 @@
-package dnd.myOcean.config.security.details.oauth2;
+package dnd.myOcean.service.sign;
 
 
+import dnd.myOcean.config.security.details.KakaoUserInfo;
 import dnd.myOcean.config.security.details.MemberDetails;
 import dnd.myOcean.domain.member.Member;
 import dnd.myOcean.repository.member.MemberRepository;
